Support fractional share quantities in trade proposals
A SELL proposal showed "SELL 1 share" of NVDD when only 0.8 shares were
held: the agent's rationale correctly described exiting the 0.8-share
position, but the proposal header and stored quantity were rounded to 1.
Approving it would try to sell more than is held.

The agent layer truncated every quantity to a whole number — the
propose_trade tool schema declared quantity as integer/minimum 1,
TradeProposal.quantity was typed int, and daily_loop cast tool inputs with
int(). The DB and broker layers already used Decimal, so fractions were
destroyed only in the agent layer.

- Tool schema and TradeProposal now carry Decimal quantities; the prompt
  tells the agent to size exits from the exact get_portfolio amount.
- daily_loop clamps every SELL to the live holding (new held_quantity
  accessor) and drops a SELL for an unheld symbol — overselling guard.
- Telegram card renders fractional quantities and strips trailing zeros.

+++ src/ai_agent/risk/rails.py
@@ -221,7 +221,7 @@ class RiskChecker:
         *,
         symbol: str,
         side: str,
-        quantity: int,
+        quantity: Decimal,
         limit_price: Decimal,
         stop_price: Decimal | None = None,
     ) -> RailResult:
